Amend the build scripts to not require existing mercurial tags

+++ utils/make_dist.py
@@ -22,12 +22,13 @@ debs, rpms, etc.
 
 Options:
     --revision=<revision>   A valid mercurial revision. An existing mercurial tag for 'release'.
-                            The revision is unsupported if a path is used as source. Uncommited
-                            changes of a path will be automatically committed and used to create the
-                            tarball. The source path is never touched, only a copy of it will be
-                            altered.
+                            The revision argument is unsupported if a path is used as argument for
+                            --source.
     --source=<source>       The source to be used. Either an URL to a Mercurial repository or local
                             path to a Mercurial repository.
+                            If a local path is used, uncommited changes will be automatically
+                            committed and used to create the tarball. The source path is never
+                            altered, only a copy of it.
                             [default: https://bitbucket.org/openattic/openattic]
     -v                      Enables the verbose mode.
     -q                      Enables the quiet mode.
@@ -122,7 +123,6 @@ class Process(object):
             if result.stderr:
                 print result.stderr
             if self.exit_on_error is True and exit_on_error is not False:
-                print 'Process exited with exit code {}.'.format(result.returncode)
                 sys.exit(result.returncode)
 
         return result
@@ -268,8 +268,8 @@ class DistBuilder(object):
 
         return latest_tag
 
-    def _determine_upcoming_version(self, revision):
-        """Determine the version using the `version.txt`.
+    def _get_version_of_revision(self, revision):
+        """Return the version of `version.txt` using the given revision.
 
         It returns the VERSION of the version.txt file in the repository. The version returned
         depends on the given revision.
@@ -299,13 +299,14 @@ class DistBuilder(object):
 
         return version
 
-    def _resolve_release_channel(self, channel):
-        """Resolve the release channel to a valid Mercurial revision.
+    def _get_revision_by_channel(self, channel):
+        """Return the revision corresponding the given release channel.
 
         This function provides the default revision for calls of this script which do NOT provide a
-        revision argument, so that the revision has to be determined automatically. If the channel
-        is 'release', the revision will resolve to the latest existing tag. If it's 'snapshot' it'll
-        resolve to the tip of the default branch.
+        revision argument, so that the revision has to be determined automatically.
+
+        If the channel is 'release', the revision will resolve to the latest existing mercurial tag.
+        If it's 'snapshot' it'll resolve to the tip of the default branch.
 
         :param channel: 'release' | 'snapshot'
         :type channel: str
@@ -409,7 +410,7 @@ class DistBuilder(object):
 
         return tag in matches if matches else False
 
-    def _get_build_basename(self, channel, revision):
+    def _get_build_basename(self, channel, version):
         """Return the base name for the given revision.
 
         Depending on the channel, this may either like `openattic-2.0.4` or
@@ -421,24 +422,8 @@ class DistBuilder(object):
         """
         assert channel in ('release', 'snapshot')
 
-        build_basename = 'openattic-'
-        if channel == 'release':
-            # revision needs to be a tag or we have to use something else here to create the
-            # basename
-            if revision is None:
-                build_basename += self._get_latest_existing_tag(strip_tag=True)
-            else:
-                if self._is_existing_tag(revision):
-                    build_basename += self._strip_mercurial_tag(revision)
-                else:
-                    # TODO Fetch the tag of `version.txt` and increment it by one?
-                    raise NotImplementedError('Currently, this feature is unsupported')
-
-        elif channel == 'snapshot':
-            if revision is None:
-                revision = self._resolve_release_channel(channel)
-            upcoming_version = self._determine_upcoming_version(revision)
-            build_basename += upcoming_version + '~' + self._datestring
+        build_basename = 'openattic-{}'.format(version)
+        build_basename += '~' + self._datestring if channel == 'snapshot' else ''
 
         return build_basename
 
@@ -561,15 +546,13 @@ class DistBuilder(object):
         self._check_dependencies(['npm'])
         self._set_npmrc_prefix()
 
-        # Delete any previous sources before retrieval but do not delete the oa_cache.
-        if isdir(self._oa_temp_build_dir):
-            rmtree(self._oa_temp_build_dir)
-
         self._retrieve_source(self._source_dir, DistBuilder.OA_CACHE_REPO_URL, skip_if_exists=True)
-        self._retrieve_source(self._source_dir, self._source)
+        if self._source_is_path and isdir(self._oa_temp_build_dir):
+            rmtree(self._oa_temp_build_dir)
+        self._retrieve_source(self._source_dir, self._source, skip_if_exists=True)
 
         channel = self._get_release_channel()
-        revision = self._args['--revision'] or self._resolve_release_channel(channel)
+        revision = self._args['--revision'] or self._get_revision_by_channel(channel)
 
         self._process.run(['hg', 'pull', '--update'], cwd=self._cache_dir)
 
@@ -582,7 +565,8 @@ class DistBuilder(object):
         for command in commands:
             self._process.run(command, cwd=self._oa_temp_build_dir)
 
-        build_basename = self._get_build_basename(channel, revision)
+        version = self._get_version_of_revision(revision)
+        build_basename = self._get_build_basename(channel, version)
         abs_tarball_file_path = self._create_source_tarball(build_basename)
         self._remove_npmrc_prefix()
